# docs/swagger.yaml
basePath: /
definitions:
  request.Acc:
    properties:
      acc:
        description: Account
        example: 1012-0000-00060776
        type: string
    type: object
  request.CID:
    properties:
      cid:
        description: Client ID
        example: 10724585
        type: integer
    type: object
  request.MultiplePaymentRequest:
    properties:
      orNumber:
        type: integer
      particulars:
        type: string
      payment: {}
      prNumber:
        type: string
      remitterCID:
        type: string
      sourceId:
        type: integer
      totalCollection:
        type: integer
      trndate:
        type: string
      username:
        type: string
    type: object
  request.OpenPaymentRequest:
    properties:
      accountNumber:
        example: "1129000022901288"
        type: string
      amount:
        example: 20
        type: integer
      particulars:
        example: Client Initiated Payment - Weekly Due Payment
        type: string
      sourceSaveAcc:
        example: 1012-0000-37622903
        type: string
      transFee:
        example: 0
        type: integer
      transFeeParticulars:
        type: string
      trnReference:
        example: CIP123-trndate01
        type: string
      username:
        example: konek2CARD
        type: string
    type: object
  request.PRNumber:
    properties:
      prNumber:
        description: PR Number
        example: AAP36467-1638243443748
        type: string
    type: object
...
  response.CustSavingInfo:
    properties:
      acc: {}
      accDesc:
        type: string
      acctType:
        type: integer
      appType:
        type: integer
      balance:
        type: number
      cid:
        type: integer
      classificationCode:
        type: integer
      classificationType:
        type: integer
      dopen:
        type: string
      status:
        type: integer
    type: object
  response.CustSavingsLists:
    properties:
      acc:
        type: string
      accDesc:
        type: string
      acctType:
        type: integer
      balance:
        type: number
      cid:
        type: integer
      dopen:
        type: string
      status:
        type: integer
      statusDesc:
        type: string
    type: object
  response.CustomerInformation:
    properties:
      accountNumber:
        type: string
      address: {}
      birthPlace:
        type: string
      brCode:
        type: integer
      business:
        type: integer
      centerCode:
        type: integer
      centerName: {}
      cid:
        type: integer
      civilStatus:
        type: integer
      classification:
        type: integer
      contact: {}
      doBirth:
        type: string
      doEntry:
        type: string
      doRecognized:
        type: string
      doResigned:
        type: string
      dosri:
        type: integer
      firstname:
        type: string
      lastname:
        type: string
      maidenFname:
        type: string
      maidenLName:
        type: string
      maidenMName:
        type: string
      memberMaidenFName:
        type: string
      memberMaidenLName:
        type: string
      memberMaidenMName:
        type: string
      middlename:
        type: string
      reffered:
        type: string
      remarks:
        type: string
      searchName: {}
      sex:
        type: integer
      status:
        type: integer
      subClassification:
        type: integer
      title:
        type: integer
      unitCode:
        type: integer
      unitName: {}
    type: object
...
  response.GenerateCol:
    properties:
      acc:
        type: string
      acctDesc:
        type: string
      accttype:
        type: integer
      amort:
        type: number
      appType:
        type: integer
      area:
        type: string
      areaCode:
        type: integer
      balInt:
        type: number
      balPrin:
        type: number
      centerCode:
        type: integer
      centerName:
        type: string
      classDesc:
        type: string
      classification:
        type: integer
      clientName:
        type: string
      code:
        type: integer
      dateStart:
        type: string
      disbdate:
        type: string
      dueInt:
        type: number
      duePrin:
        type: number
      gives:
        type: integer
      ibalInt:
        type: number
      ibalPrin:
        type: number
      iiid:
        type: integer
      interest:
        type: number
      loanBal:
        type: number
      manCode:
        type: integer
      maturity:
        type: string
      pay:
        type: number
      principal:
        type: number
      saveBal:
        type: number
      staffName:
        type: string
      status:
        type: integer
      statusDesc:
        type: string
      type:
        type: integer
      um:
        type: string
      unit:
        type: string
      unpaidCtr:
        type: integer
      uuid: {}
      waiveInt:
        type: number
      withdraw:
        type: number
      writeoff:
        type: integer
      writtenoff:
        type: integer
    type: object
  response.LoanInfo:
    properties:
      acc:
        type: string
      accDesc:
        type: string
      acctType:
        type: integer
      amort: {}
      amortList:
        items:
          $ref: '#/definitions/response.AmortList'
        type: array
      amortcond:
        type: integer
      amortcondvalue:
        type: number
      annumdiv:
        type: integer
      appType:
        type: integer
      approvby:
        type: string
      balance:
        type: number
      charges: {}
      cid:
        type: integer
      classification_code:
        type: integer
      classification_type:
        type: integer
      conintrate:
        type: number
      cycle:
        type: integer
      disbby:
        type: string
      discounted:
        type: number
      domaturity:
        type: string
      dopen:
        type: string
      frequency:
        type: integer
      fundsource:
        type: integer
      interest:
        type: number
      intr:
        type: number
      isLumpsum:
        type: integer
      lngrpcode:
        type: integer
      loanID: {}
      netproceed:
        type: number
      oth:
        type: number
      others:
        type: number
      penalty:
        type: number
      prin:
        type: number
      principal:
        type: number
      proff:
        type: integer
      remarks: {}
      status:
        type: integer
      term:
        type: integer
      waivedint:
        type: number
      weekspaid:
        type: integer
    type: object
...
  response.OpenPaymentResponse:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/response.Details'
        type: array
      message:
        type: string
    type: object
  response.ResponseModel:
    properties:
      data: {}
      message:
        type: string
      retCode:
        type: string
    type: object
  response.ResponseModelwoData:
    properties:
      message:
        type: string
      retCode:
        type: integer
    type: object
  response.SuperSaving:
    properties:
      acc:
        type: string
      balance:
        type: number
      centerCode:
        type: integer
      centerName:
        type: string
      cid:
        type: integer
      fullname:
        type: string
      unitCode:
        type: integer
      unitName:
        type: string
      withdrawable:
        type: number
      withdrawalAmount:
        type: number
    type: object
host: 192.168.0.137:5000
info:
  contact:
    name: FDSAP Support
  description: JANUS Swagger JanusGate
  license: {}
  termsOfService: http://swagger.io/terms/
  title: JANUS V2
  version: "2.0"
paths:
  /api/public/v1/coremicro/callBackReference:
    post:
      consumes:
      - application/json
      description: Used to check if the transaction is posted in CORE
      parameters:
      - description: Reference Number
        in: body
        name: prNumber
        required: true
        schema:
          $ref: '#/definitions/request.PRNumber'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseModelwoData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-TRANSACTION-CHECKER
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/custSavingInfo:
    post:
      consumes:
      - application/json
      description: Used to get the details of savings of NGO clients
      parameters:
      - description: Account Number
        in: body
        name: acc
        required: true
        schema:
          $ref: '#/definitions/request.Acc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CustSavingInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-SAVINGS-INFORMATION
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/customerSavingsList/:
    post:
      consumes:
      - application/json
      description: Used to get the list of savings of NGO clients
      parameters:
      - description: Client ID
        in: body
        name: cid
        required: true
        schema:
          $ref: '#/definitions/request.CID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CustSavingsLists'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-CLIENT-SAVINGS-LIST
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/generateColCid/:
    post:
      consumes:
      - application/json
      description: Used to get the weekly due of the clients
      parameters:
      - description: Client ID
        in: body
        name: cid
        required: true
        schema:
          $ref: '#/definitions/request.CID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.GenerateCol'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-DUE-GENERATOR
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/getSavingsForSuperApp:
    post:
      consumes:
      - application/json
      description: Used to get the total and available amount of client savings
      parameters:
      - description: Client ID
        in: body
        name: cid
        required: true
        schema:
          $ref: '#/definitions/request.CID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuperSaving'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-AVAILABLE-BALANCE
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/loanInfo:
    post:
      consumes:
      - application/json
      description: Used to get the information of loan and amortization
      parameters:
      - description: Account Number
        in: body
        name: acc
        required: true
        schema:
          $ref: '#/definitions/request.Acc'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoanInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-LOAN-INFORMATION
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/multiplePayment:
    post:
      consumes:
      - application/json
      description: Used to get the information of loan and amortization
      parameters:
      - description: Payment Request
        in: body
        name: paymentRequest
        required: true
        schema:
          $ref: '#/definitions/request.MultiplePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseModelwoData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-LOAN-INFORMATION
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/openPaymentTransaction:
    post:
      consumes:
      - application/json
      description: Old API used to pay loan in CORE
      parameters:
      - description: Search
        in: body
        name: paymentRequest
        required: true
        schema:
          $ref: '#/definitions/request.OpenPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.OpenPaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-OLD-LOAN-PAYMENT
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/searchCustomerCID:
    post:
      consumes:
      - application/json
      description: Used to get the details of NGO clients
      parameters:
      - description: Client ID
        in: body
        name: cid
        required: true
        schema:
          $ref: '#/definitions/request.CID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CustomerInformation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-CLIENT-INFORMATION
      tags:
      - CORE MICRO
  /api/public/v1/coremicro/searchLoanList:
    post:
      consumes:
      - application/json
      description: Used to get the details of NGO clients
      parameters:
      - description: Client ID
        in: body
        name: cid
        required: true
        schema:
          $ref: '#/definitions/request.CID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.LoanList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ResponseModel'
      summary: CORE-CLIENT-INFORMATION
      tags:
      - CORE MICRO
swagger: "2.0"